Move COGLayers test code into separate Specs
